requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.1
attrs==22.2.0
awsebcli==3.20.10
backcall==0.2.0
bcrypt==4.0.1
blessed==1.20.0
boto3==1.26.111
botocore==1.29.111
Brotli==1.0.9
cached-property==1.5.2
cairocffi==1.5.1
CairoSVG==2.7.0
cement==2.8.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
comm==0.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
cssselect2==0.7.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.3
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-extra-fields==3.0.2
django-js-asset==2.0.0
django-nested-admin==4.0.2
django-rest-framework==0.1.0
django-ses==3.3.0
django-shell-plus==1.1.7
django-storages==1.14.4
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docker==4.4.4
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.1.0
executing==1.2.0
fonttools==4.39.0
html5lib==1.1
idna==3.4
importlib-metadata==6.3.0
ipython==8.11.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
lxml==5.2.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy
oauthlib==3.2.2
odfpy==1.4.1
openai==0.28.0
openpyxl==3.1.5
packaging==23.0
pandas
pangocffi==0.11.0
paramiko==3.1.0
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pipdeptree==2.7.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2-binary
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydyf==0.11.0
Pygments==2.14.0
PyJWT==2.6.0
PyNaCl==1.5.0
PyPDF2==3.0.1
pyphen==0.13.2
pyrsistent==0.19.3
python-dotenv==1.0.1
python-dateutil==2.8.2
python-docx==1.1.0
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2022.6
PyYAML
pyzmq==25.0.1
reportlab==4.0.4
requests==2.32.3
requests-oauthlib==1.3.1
s3transfer==0.6.0
semantic-version==2.8.5
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stack-data==0.6.2
stripe==10.7.0
termcolor==1.1.0
texttable==1.6.7
tinycss2==1.3.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.1.9
weasyprint==52.5
webencodings==0.5.1
websocket-client==0.59.0
whitenoise==6.4.0
zipp
zopfli